Fix bug at Home not showing correct date/games - closes #51
Sometimes the game starts in a day but ends in the next day,
if we're in the next day already but the game started before,
it does not show the results.

In order to fix this, I'm using the timezone from the games (America/New_York)
and creating a "span" of 6h to show the correct games that are happening.